update function to use internet and binary pkgs
This adds the option to download and install packages to the repo
The only limitation it has right now is the fact that it does nothing
to check for dependencies.

This employs a bunch of try statements so that it will attempt to blow
through functions that may return errors to get to the end of the
process.